DELETE request
Deleted a document in the database using a delete request. We did this by sending a delete request from the front end using the 'fetch API'. This request is sent in AJAX form. And the we handle the delete request to the route in the backend. Deleted the document using '.findByIdAndDelete' method in mongoose. And configured a redirect to home method. As the request was in AJAX form we are not able to use the 'redirect()' method directly so we sent a JSON response to the browser and configured the redirect using that response in the frontend